Deduplicate common arguments into a concern
