+++ lib/rails_template18f/version.rb
