+++ lib/rails_template_18f.rb
@@ -1,8 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "rails_template_18f/version"
+require_relative "rails_template18f/version"
 
 module RailsTemplate18f
+  extend ActiveSupport::Autoload
+
+  autoload :TerraformOptions
+
   class Error < StandardError; end
 
   class Railtie < ::Rails::Railtie; end
